Refactor UI components, support color theme changing.
