+++ src/main/java/org/apache/lucene/luke/app/desktop/Preferences.java
@@ -17,6 +17,7 @@
 
 package org.apache.lucene.luke.app.desktop;
 
+import java.awt.Color;
 import java.io.IOException;
 import java.util.List;
 
@@ -37,4 +38,30 @@ public interface Preferences {
   boolean isKeepAllCommits();
 
   void setIndexOpenerPrefs(boolean readOnly, String dirImpl, boolean noReader, boolean useCompound, boolean keepAllCommits) throws IOException;
+
+  ColorTheme getColorTheme();
+
+  void setColorTheme(ColorTheme theme) throws IOException;
+
+  enum ColorTheme {
+
+    /* Gray theme */
+    GRAY(Color.decode("#e6e6e6")),
+    /* Classic theme */
+    CLASSIC(Color.decode("#ece9d0")),
+    /* Sandstone theme */
+    SANDSTONE(Color.decode("#ddd9d4")),
+    /* Navy theme */
+    NAVY(Color.decode("#e6e6ff"));
+
+    private Color backgroundColor;
+
+    ColorTheme(Color backgroundColor) {
+      this.backgroundColor = backgroundColor;
+    }
+
+    public Color getBackgroundColor() {
+      return backgroundColor;
+    }
+  }
 }
