+++ src/main/java/org/apache/lucene/luke/app/desktop/components/AnalysisPanelProvider.java
@@ -132,9 +132,11 @@ public final class AnalysisPanelProvider implements Provider<JPanel>, AnalysisTa
   @Override
   public JPanel get() {
     JPanel panel = new JPanel(new GridLayout(1, 1));
+    panel.setOpaque(false);
     panel.setBorder(BorderFactory.createLineBorder(Color.gray));
 
     JSplitPane splitPane = new JSplitPane(JSplitPane.VERTICAL_SPLIT, initUpperPanel(), initLowerPanel());
+    splitPane.setOpaque(false);
     splitPane.setDividerLocation(320);
     panel.add(splitPane);
 
@@ -142,6 +144,7 @@ public final class AnalysisPanelProvider implements Provider<JPanel>, AnalysisTa
   }
 
   private JPanel initUpperPanel() {
+    mainPanel.setOpaque(false);
     mainPanel.setLayout(new BorderLayout());
     mainPanel.setBorder(BorderFactory.createEmptyBorder(3, 3, 3, 3));
 
@@ -153,6 +156,7 @@ public final class AnalysisPanelProvider implements Provider<JPanel>, AnalysisTa
 
   private JPanel initSwitcherPanel() {
     JPanel panel = new JPanel(new FlowLayout(FlowLayout.LEADING));
+    panel.setOpaque(false);
 
     presetRB.setText(MessageUtils.getLocalizedMessage("analysis.radio.preset"));
     presetRB.setActionCommand(TYPE_PRESET);
@@ -176,8 +180,10 @@ public final class AnalysisPanelProvider implements Provider<JPanel>, AnalysisTa
 
   private JPanel initLowerPanel() {
     JPanel inner1 = new JPanel(new BorderLayout());
+    inner1.setOpaque(false);
 
     JPanel analyzerName = new JPanel(new FlowLayout(FlowLayout.LEADING, 10, 2));
+    analyzerName.setOpaque(false);
     analyzerName.add(new JLabel(MessageUtils.getLocalizedMessage("analysis.label.selected_analyzer")));
     analyzerNameLbl.setText(analysisModel.currentAnalyzer().getClass().getName());
     analyzerName.add(analyzerNameLbl);
@@ -193,6 +199,7 @@ public final class AnalysisPanelProvider implements Provider<JPanel>, AnalysisTa
     inner1.add(analyzerName, BorderLayout.PAGE_START);
 
     JPanel input = new JPanel(new FlowLayout(FlowLayout.LEADING, 5, 2));
+    input.setOpaque(false);
     inputArea.setRows(3);
     inputArea.setColumns(50);
     inputArea.setLineWrap(true);
@@ -221,8 +228,10 @@ public final class AnalysisPanelProvider implements Provider<JPanel>, AnalysisTa
     inner1.add(input, BorderLayout.CENTER);
 
     JPanel inner2 = new JPanel(new BorderLayout());
+    inner2.setOpaque(false);
 
     JPanel hint = new JPanel(new FlowLayout(FlowLayout.LEADING));
+    hint.setOpaque(false);
     hint.add(new JLabel(MessageUtils.getLocalizedMessage("analysis.hint.show_attributes")));
     inner2.add(hint, BorderLayout.PAGE_START);
 
@@ -239,6 +248,7 @@ public final class AnalysisPanelProvider implements Provider<JPanel>, AnalysisTa
     inner2.add(new JScrollPane(tokensTable), BorderLayout.CENTER);
 
     JPanel panel = new JPanel(new BorderLayout());
+    panel.setOpaque(false);
     panel.setBorder(BorderFactory.createEmptyBorder(3, 3, 3, 3));
     panel.add(inner1, BorderLayout.PAGE_START);
     panel.add(inner2, BorderLayout.CENTER);
