+++ src/main/java/org/apache/lucene/luke/app/desktop/components/DocumentsPanelProvider.java
@@ -189,9 +189,11 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
   @Override
   public JPanel get() {
     JPanel panel = new JPanel(new GridLayout(1, 1));
+    panel.setOpaque(false);
     panel.setBorder(BorderFactory.createLineBorder(Color.gray));
 
     JSplitPane splitPane = new JSplitPane(JSplitPane.VERTICAL_SPLIT, initUpperPanel(), initLowerPanel());
+    splitPane.setOpaque(false);
     splitPane.setDividerLocation(0.4);
     panel.add(splitPane);
 
@@ -202,6 +204,7 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
 
   private JPanel initUpperPanel() {
     JPanel panel = new JPanel(new GridBagLayout());
+    panel.setOpaque(false);
     GridBagConstraints c = new GridBagConstraints();
 
     c.gridx = 0;
@@ -223,15 +226,18 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
 
   private JPanel initBrowseTermsPanel() {
     JPanel panel = new JPanel(new BorderLayout());
+    panel.setOpaque(false);
     panel.setBorder(BorderFactory.createEmptyBorder(3, 3, 3, 3));
 
     JPanel top = new JPanel(new FlowLayout(FlowLayout.LEADING));
+    top.setOpaque(false);
     JLabel label = new JLabel(MessageUtils.getLocalizedMessage("documents.label.browse_terms"));
     top.add(label);
 
     panel.add(top, BorderLayout.PAGE_START);
 
     JPanel center = new JPanel(new GridBagLayout());
+    center.setOpaque(false);
     GridBagConstraints c = new GridBagConstraints();
     c.fill = GridBagConstraints.BOTH;
 
@@ -276,6 +282,7 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
     panel.add(center, BorderLayout.CENTER);
 
     JPanel footer = new JPanel(new FlowLayout(FlowLayout.LEADING, 20, 5));
+    footer.setOpaque(false);
     JLabel hintLbl = new JLabel(MessageUtils.getLocalizedMessage("documents.label.browse_terms_hint"));
     footer.add(hintLbl);
     panel.add(footer, BorderLayout.PAGE_END);
@@ -285,9 +292,11 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
 
   private JPanel initBrowseDocsByTermPanel() {
     JPanel panel = new JPanel(new BorderLayout());
+    panel.setOpaque(false);
     panel.setBorder(BorderFactory.createEmptyBorder(3, 3, 3, 3));
 
     JPanel center = new JPanel(new GridBagLayout());
+    center.setOpaque(false);
     GridBagConstraints c = new GridBagConstraints();
     c.fill = GridBagConstraints.BOTH;
 
@@ -362,17 +371,21 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
 
   private JPanel initLowerPanel() {
     JPanel panel = new JPanel(new BorderLayout());
+    panel.setOpaque(false);
     panel.setBorder(BorderFactory.createEmptyBorder(3, 3, 3, 3));
 
     JPanel browseDocsPanel = new JPanel();
+    browseDocsPanel.setOpaque(false);
     browseDocsPanel.setLayout(new BoxLayout(browseDocsPanel, BoxLayout.PAGE_AXIS));
     browseDocsPanel.add(initBrowseDocsBar());
 
     JPanel browseDocsNote1 = new JPanel(new FlowLayout(FlowLayout.LEADING));
+    browseDocsNote1.setOpaque(false);
     browseDocsNote1.add(new JLabel(MessageUtils.getLocalizedMessage("documents.label.doc_table_note1")));
     browseDocsPanel.add(browseDocsNote1);
 
     JPanel browseDocsNote2 = new JPanel(new FlowLayout(FlowLayout.LEADING));
+    browseDocsNote2.setOpaque(false);
     browseDocsNote2.add(new JLabel(MessageUtils.getLocalizedMessage("documents.label.doc_table_note2")));
     browseDocsPanel.add(browseDocsNote2);
 
@@ -389,6 +402,7 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
         DocumentsTableModel.Column.NORM.getColumnWidth(),
         DocumentsTableModel.Column.VALUE.getColumnWidth());
     JPanel flagsHeader = new JPanel(new FlowLayout(FlowLayout.CENTER));
+    flagsHeader.setOpaque(false);
     flagsHeader.add(new JLabel("Flags"));
     flagsHeader.add(new JLabel("Help"));
     documentTable.getColumnModel().getColumn(DocumentsTableModel.Column.FLAGS.getIndex()).setHeaderValue(flagsHeader);
@@ -402,9 +416,11 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
 
   private JPanel initBrowseDocsBar() {
     JPanel panel = new JPanel(new GridLayout(1, 2));
+    panel.setOpaque(false);
     panel.setBorder(BorderFactory.createEmptyBorder(5, 0, 0, 5));
 
     JPanel left = new JPanel(new FlowLayout(FlowLayout.LEADING, 10, 2));
+    left.setOpaque(false);
     JLabel label = new JLabel(FontUtils.elegantIconHtml("&#x68;", MessageUtils.getLocalizedMessage("documents.label.browse_doc_by_idx")));
     label.setHorizontalTextPosition(JLabel.LEFT);
     left.add(label);
@@ -416,6 +432,7 @@ public final class DocumentsPanelProvider implements Provider<JPanel>, Documents
     panel.add(left);
 
     JPanel right = new JPanel(new FlowLayout(FlowLayout.TRAILING));
+    right.setOpaque(false);
     copyDocValuesBtn.setText(FontUtils.elegantIconHtml("&#xe0e6;", MessageUtils.getLocalizedMessage("documents.buttont.copy_values")));
     copyDocValuesBtn.setMargin(new Insets(5, 0, 5, 0));
     copyDocValuesBtn.addActionListener(listeners::copySelectedOrAllStoredValues);
