+++ src/main/java/org/apache/lucene/luke/app/desktop/components/dialog/ConfirmDialogFactory.java
@@ -18,78 +18,10 @@
 package org.apache.lucene.luke.app.desktop.components.dialog;
 
 import org.apache.lucene.luke.app.desktop.util.DialogOpener;
-import org.apache.lucene.luke.app.desktop.util.FontUtils;
-import org.apache.lucene.luke.app.desktop.util.MessageUtils;
 import org.apache.lucene.luke.app.desktop.util.lang.Callable;
 
-import javax.swing.BorderFactory;
-import javax.swing.JButton;
-import javax.swing.JDialog;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import java.awt.BorderLayout;
-import java.awt.Color;
-import java.awt.Dialog;
-import java.awt.Dimension;
-import java.awt.FlowLayout;
-import java.awt.Font;
-import java.awt.GridLayout;
-import java.awt.Window;
+public interface ConfirmDialogFactory extends DialogOpener.DialogFactory {
+  void setMessage(String message);
 
-public final class ConfirmDialogFactory implements DialogOpener.DialogFactory {
-
-  private JDialog dialog;
-
-  private String message;
-
-  private Callable callback;
-
-  public void setMessage(String message) {
-    this.message = message;
-  }
-
-  public void setCallback(Callable callback) {
-    this.callback = callback;
-  }
-
-  @Override
-  public JDialog create(Window owner, String title, int width, int height) {
-    dialog = new JDialog(owner, title, Dialog.ModalityType.APPLICATION_MODAL);
-    dialog.add(content());
-    dialog.setSize(new Dimension(width, height));
-    dialog.setLocationRelativeTo(owner);
-    return dialog;
-  }
-
-  private JPanel content() {
-    JPanel panel = new JPanel(new BorderLayout());
-    panel.setBorder(BorderFactory.createEmptyBorder(15, 15, 15, 15));
-
-    JPanel header = new JPanel(new FlowLayout(FlowLayout.LEADING));
-    JLabel alertIconLbl = new JLabel(FontUtils.elegantIconHtml("&#x71;"));
-    alertIconLbl.setHorizontalAlignment(JLabel.CENTER);
-    alertIconLbl.setFont(new Font(alertIconLbl.getFont().getFontName(), Font.PLAIN, 25));
-    header.add(alertIconLbl);
-    panel.add(header, BorderLayout.PAGE_START);
-
-    JPanel center = new JPanel(new GridLayout(1, 1));
-    center.setBorder(BorderFactory.createLineBorder(Color.gray, 3));
-    center.add(new JLabel(message, JLabel.CENTER));
-    panel.add(center, BorderLayout.CENTER);
-
-    JPanel footer = new JPanel(new FlowLayout(FlowLayout.TRAILING));
-    JButton okBtn = new JButton(MessageUtils.getLocalizedMessage("button.ok"));
-    okBtn.addActionListener(e -> {
-      callback.call();
-      dialog.dispose();
-    });
-    footer.add(okBtn);
-    JButton closeBtn = new JButton(MessageUtils.getLocalizedMessage("button.close"));
-    closeBtn.addActionListener(e -> dialog.dispose());
-    footer.add(closeBtn);
-    panel.add(footer, BorderLayout.PAGE_END);
-
-    return panel;
-  }
-
+  void setCallback(Callable callback);
 }
