+++ src/main/java/org/apache/lucene/luke/app/desktop/components/dialog/analysis/AnalysisChainDialogFactory.java
@@ -19,117 +19,7 @@ package org.apache.lucene.luke.app.desktop.components.dialog.analysis;
 
 import org.apache.lucene.analysis.custom.CustomAnalyzer;
 import org.apache.lucene.luke.app.desktop.util.DialogOpener;
-import org.apache.lucene.luke.app.desktop.util.MessageUtils;
 
-import javax.swing.BorderFactory;
-import javax.swing.JButton;
-import javax.swing.JDialog;
-import javax.swing.JLabel;
-import javax.swing.JList;
-import javax.swing.JPanel;
-import javax.swing.JScrollPane;
-import javax.swing.JTextField;
-import java.awt.BorderLayout;
-import java.awt.Color;
-import java.awt.Dialog;
-import java.awt.Dimension;
-import java.awt.FlowLayout;
-import java.awt.GridBagConstraints;
-import java.awt.GridBagLayout;
-import java.awt.Insets;
-import java.awt.Window;
-
-public class AnalysisChainDialogFactory implements DialogOpener.DialogFactory {
-
-  private JDialog dialog;
-
-  private CustomAnalyzer analyzer;
-
-  public void setAnalyzer(CustomAnalyzer analyzer) {
-    this.analyzer = analyzer;
-  }
-
-  @Override
-  public JDialog create(Window owner, String title, int width, int height) {
-    dialog = new JDialog(owner, title, Dialog.ModalityType.APPLICATION_MODAL);
-    dialog.add(content());
-    dialog.setSize(new Dimension(width, height));
-    dialog.setLocationRelativeTo(owner);
-    return dialog;
-  }
-
-  private JPanel content() {
-    JPanel panel = new JPanel(new BorderLayout());
-    panel.setBorder(BorderFactory.createEmptyBorder(10, 10, 10, 10));
-
-    panel.add(analysisChain(), BorderLayout.PAGE_START);
-
-    JPanel footer = new JPanel(new FlowLayout(FlowLayout.TRAILING, 10, 5));
-    JButton closeBtn = new JButton(MessageUtils.getLocalizedMessage("button.close"));
-    closeBtn.addActionListener(e -> dialog.dispose());
-    footer.add(closeBtn);
-    panel.add(footer, BorderLayout.PAGE_END);
-
-    return panel;
-  }
-
-  private JPanel analysisChain() {
-    JPanel panel = new JPanel(new GridBagLayout());
-
-    GridBagConstraints c = new GridBagConstraints();
-    c.fill = GridBagConstraints.HORIZONTAL;
-    c.insets = new Insets(5, 5, 5, 5);
-
-    c.gridx = 0;
-    c.gridy = 0;
-    c.weightx = 0.1;
-    c.weighty = 0.5;
-    panel.add(new JLabel(MessageUtils.getLocalizedMessage("analysis.dialog.chain.label.charfilters")), c);
-
-    String[] charFilters = analyzer.getCharFilterFactories().stream().map(f -> f.getClass().getName()).toArray(String[]::new);
-    JList<String> charFilterList = new JList<>(charFilters);
-    charFilterList.setVisibleRowCount(charFilters.length == 0 ? 1 : Math.min(charFilters.length, 5));
-    c.gridx = 1;
-    c.gridy = 0;
-    c.weightx = 0.5;
-    c.weighty = 0.5;
-    panel.add(new JScrollPane(charFilterList), c);
-
-    c.gridx = 0;
-    c.gridy = 1;
-    c.weightx = 0.1;
-    c.weighty = 0.1;
-    panel.add(new JLabel(MessageUtils.getLocalizedMessage("analysis.dialog.chain.label.tokenizer")), c);
-
-    String tokenizer = analyzer.getTokenizerFactory().getClass().getName();
-    JTextField tokenizerTF = new JTextField(tokenizer);
-    tokenizerTF.setColumns(30);
-    tokenizerTF.setEditable(false);
-    tokenizerTF.setPreferredSize(new Dimension(300, 25));
-    tokenizerTF.setBorder(BorderFactory.createLineBorder(Color.gray));
-    c.gridx = 1;
-    c.gridy = 1;
-    c.weightx = 0.5;
-    c.weighty = 0.1;
-    panel.add(tokenizerTF, c);
-
-    c.gridx = 0;
-    c.gridy = 2;
-    c.weightx = 0.1;
-    c.weighty = 0.5;
-    panel.add(new JLabel(MessageUtils.getLocalizedMessage("analysis.dialog.chain.label.tokenfilters")), c);
-
-    String[] tokenFilters = analyzer.getTokenFilterFactories().stream().map(f -> f.getClass().getName()).toArray(String[]::new);
-    JList<String> tokenFilterList = new JList<>(tokenFilters);
-    tokenFilterList.setVisibleRowCount(tokenFilters.length == 0 ? 1 : Math.min(tokenFilters.length, 5));
-    tokenFilterList.setMinimumSize(new Dimension(300, 25));
-    c.gridx = 1;
-    c.gridy = 2;
-    c.weightx = 0.5;
-    c.weighty = 0.5;
-    panel.add(new JScrollPane(tokenFilterList), c);
-
-    return panel;
-  }
-
+public interface AnalysisChainDialogFactory extends DialogOpener.DialogFactory {
+  void setAnalyzer(CustomAnalyzer analyzer);
 }
