+++ src/main/resources/messages.properties
@@ -186,7 +186,7 @@ analysis_custom.label.tokenfilters=Token Filters
 analysis_custom.label.selected=Selected
 analysis_custom.label.add=Add
 analysis_custom.label.set=Set
-analysis_custom.label.edit=Edit
+analysis_custom.label.edit=Show & Edit
 # Search
 search.label.settings=Query settings
 search.label.expression=Query expression
